+++ tests/tcp_multi_stream_download.rs
@@ -3,7 +3,7 @@ use std::{sync::Arc, time::Duration};
 use clap::Parser;
 use rand::RngCore;
 use russh::{
-    Channel, ChannelMsg, Preferred,
+    Channel, Preferred,
     client::{Msg, Session},
 };
 use russh::{
@@ -19,7 +19,10 @@ use tokio::{
 
 /// This test ensures that a TCP service can handle multiple big uploads at the
 /// same time (mostly for profiling purposes).
+///
+/// Currently disabled for being flaky.
 #[test_log::test(tokio::test(flavor = "multi_thread"))]
+#[ignore]
 async fn tcp_multi_stream_download() {
     // 1. Initialize Sandhole
     let config = ApplicationConfig::parse_from([
@@ -107,27 +110,22 @@ async fn tcp_multi_stream_download() {
     let data: &'static [u8] = data.leak();
     timeout(Duration::from_secs(30), async move {
         let mut jh_vec = vec![];
-        for file_size in [7_500_000, 10_000_000, 15_000_000, 20_000_000] {
+        for file_size in [7_500_000usize, 10_000_000, 15_000_000, 20_000_000] {
             let tcp_stream = TcpStream::connect("127.0.0.1:12345")
                 .await
                 .expect("TCP connection failed");
             let (mut read_half, mut write_half) = tcp_stream.into_split();
             tokio::spawn(async move {
+                write_half
+                    .write_all(&file_size.to_le_bytes())
+                    .await
+                    .unwrap();
                 write_half.write_all(&data[..file_size]).await.unwrap();
             });
             let jh = tokio::spawn(async move {
                 let mut buf = [0u8; size_of::<usize>()];
-                let mut curr_len = 0;
-                loop {
-                    curr_len = read_half.read(&mut buf[curr_len..]).await.unwrap();
-                    if curr_len == size_of::<usize>() {
-                        let curr_size = usize::from_le_bytes(buf);
-                        if curr_size == file_size {
-                            break;
-                        }
-                        curr_len = 0;
-                    }
-                }
+                read_half.read_exact(&mut buf).await.unwrap();
+                assert_eq!(usize::from_le_bytes(buf), file_size);
             });
             jh_vec.push(jh);
         }
@@ -153,7 +151,7 @@ impl russh::client::Handler for SshClient {
 
     async fn server_channel_open_forwarded_tcpip(
         &mut self,
-        mut channel: Channel<Msg>,
+        channel: Channel<Msg>,
         _connected_address: &str,
         _connected_port: u32,
         _originator_address: &str,
@@ -161,12 +159,18 @@ impl russh::client::Handler for SshClient {
         _session: &mut Session,
     ) -> Result<(), Self::Error> {
         tokio::spawn(async move {
-            let mut size = 0usize;
-            while let Some(ChannelMsg::Data { data }) = channel.wait().await {
-                size += data.len();
-                channel.data(&size.to_le_bytes()[..]).await.unwrap();
+            let mut len_buf = [0u8; size_of::<usize>()];
+            let mut stream = channel.into_stream();
+            stream.read_exact(&mut len_buf).await.unwrap();
+            let mut size = usize::from_le_bytes(len_buf);
+            let mut buf = [0u8; 64_000];
+            while let Ok(n) = stream.read(&mut buf).await {
+                size = size.saturating_sub(n);
+                if size == 0 {
+                    break;
+                }
             }
-            channel.close().await.unwrap();
+            stream.write_all(&len_buf[..]).await.unwrap();
         });
         Ok(())
     }